fix dump v3
