+++ meilisearch-lib/src/index_controller/dump_actor/loaders/mod.rs
@@ -1,2 +1,3 @@
 pub mod v1;
-pub mod v2;
+pub mod v3;
+
